Rolins Fix

+++ src/main/java/app/controllers/AdminController.java
@@ -8,7 +8,6 @@ import app.persistence.ConnectionPool;
 import app.persistence.UserMapper;
 import io.javalin.http.Context;
 
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,7 +17,7 @@ public class AdminController {
         try{
             User user = ctx.sessionAttribute("chosen_user");
             if(AdminMapper.deleteUser(user.getEmail(), connectionPool)){
-                ctx.attribute("message","User: "+user.getEmail()+" Has been deleted");
+                ctx.attribute("message","User: " + user.getEmail() + " Has been deleted");
                 ctx.render("admin.html");
             }
         }catch (DatabaseException | NullPointerException e){
@@ -28,35 +27,47 @@ public class AdminController {
     }
 
     public static void deleteAnUserByUsingUserIdAndOrderId(Context ctx, ConnectionPool connectionPool) throws DatabaseException{
+        checkUsers(ctx, connectionPool);
         String userEmail = ctx.sessionAttribute("userEmail");
-        int orderid = Integer.parseInt(ctx.formParam("selectedOrder"));
         try {
+            if(ctx.formParam("selected_order") == null){
+                throw new AssertionError();
+            }
+            int orderid = Integer.parseInt(ctx.formParam("selected_order"));
             UserMapper.deleteUserSpecificOrder(orderid,connectionPool);
-            ctx.render("admin.html");
-        }catch (DatabaseException e){
-            System.out.println(e);
-            ctx.render("admin.html");
+            List<Order> orderlist = ctx.sessionAttribute("orderlist");
+            assert orderlist != null;
+            for(int i = 0; i < orderlist.size(); i++){
+                if(orderlist.get(i).getOrderId()==orderid){
+                    orderlist.remove(i);
+                    break;
+                }
+            }
+            ctx.render("admin.html"); // TODO
+        } catch (DatabaseException|AssertionError e){
+            ctx.redirect("/adminInfoForUsers");
+            ctx.attribute("message",e.getMessage());
         }
 
     }
 
-    public static void updateUserBalanceUsingEmail(Context ctx, ConnectionPool connectionPool) throws DatabaseException{
-       try {
-        String userEmail = ctx.formParam("userEmail");
-        int balanceUpdateAmount = Integer.parseInt(ctx.formParam("updateBalance"));
-        int newbalance = UserMapper.getBalanceViaEmail(userEmail,connectionPool) +balanceUpdateAmount;
-        AdminMapper.updateBalance(userEmail,newbalance,connectionPool);
-        ctx.attribute("message", userEmail+"'s balance has successfully been updated with: "+balanceUpdateAmount+" and is now: "+newbalance);
-        ctx.render("admin.html");
-    }catch (DatabaseException| NullPointerException e){
-        ctx.attribute("message", e.getMessage());
-           System.out.println(e);
-           ctx.render("admin.html");
-    }catch (NumberFormatException e){
-        ctx.attribute("message","Error, make sure you enter a number");
-        ctx.render("admin.html"); // TODO
-    }
-
+    public static void updateUserBalanceUsingEmail(Context ctx, ConnectionPool connectionPool) throws DatabaseException {
+        checkUsers(ctx, connectionPool);
+        try {
+            String userEmail = ctx.formParam("userEmail");
+            int balanceUpdateAmount = Integer.parseInt(ctx.formParam("updateBalance"));
+            int newbalance = UserMapper.getBalanceViaEmail(userEmail, connectionPool) + balanceUpdateAmount;
+            AdminMapper.updateBalance(userEmail,newbalance,connectionPool);
+            ctx.attribute("message", userEmail + "'s balance has successfully been updated with: " + balanceUpdateAmount + " and is now: " + newbalance);
+            ctx.render("admin.html");
+        }catch (DatabaseException| NullPointerException e){
+            ctx.attribute("message", e.getMessage());
+               System.out.println(e);
+               ctx.render("admin.html");
+        }catch (NumberFormatException e){
+            ctx.attribute("message","Error, make sure you enter a number");
+            ctx.render("admin.html"); // TODO
+        }
     }
     public static void updateUserBalance(Context ctx, ConnectionPool connectionPool){
         try {
@@ -92,13 +103,8 @@ public class AdminController {
         }
     }
     public static void allUsers(Context ctx, ConnectionPool connectionPool) {
-        List<User> userList;
         try {
-            ctx.sessionAttribute("admin_chosen_user", new User());
-            userList = AdminMapper.getAllUsers(connectionPool);
-            //System.out.println("Userlist: " + userList);
-            //ctx.attribute("userlist", userList);
-            ctx.sessionAttribute("userlist",userList);
+            checkUsers(ctx, connectionPool);
             ctx.render("admin.html");        // TODO
         } catch (DatabaseException e) {
             ctx.attribute("message", e.getMessage());
@@ -106,30 +112,55 @@ public class AdminController {
         }
     }
 
+    public static void checkUsers(Context ctx, ConnectionPool connectionPool) throws DatabaseException{
+        if(ctx.sessionAttribute("selected_order")==null){
+            getUsers(ctx, connectionPool);
+        }
+    }
+
+    public static void getUsers(Context ctx, ConnectionPool connectionPool) throws DatabaseException{
+        List<User> userList;
+        userList = AdminMapper.getAllUsers(connectionPool);
+        ctx.sessionAttribute("userlist",userList);
+        if(ctx.formParam("selected_user")!=null){
+            getAllOrdersFromCostumer(ctx, connectionPool);
+        }
+    }
 
 
-    public static void getAllOrdersFromCostumers(Context ctx, ConnectionPool connectionPool){
+
+    public static void getAllOrdersFromCostumer(Context ctx, ConnectionPool connectionPool){
         List<Order> orderList;
         try {
-            //int userID = Integer.parseInt(ctx.formParam("selected_user"));
             String userEmail = ctx.formParam("selected_user");
             ctx.sessionAttribute("userEmail",userEmail);
             User user = UserMapper.getUserByEmail(userEmail,connectionPool);
             System.out.println(user);
             ctx.sessionAttribute("admin_chosen_user",user);
-            orderList = AdminMapper.getAllOrdersFromCostumer(user.getUserID(),connectionPool);
+            if(user!=null){
+                orderList = AdminMapper.getAllOrdersFromCostumer(user.getUserID(), connectionPool);
+            } else {
+                orderList = null;
+            }
             ctx.sessionAttribute("orderlist",orderList);
             ctx.render("admin.html");
         }catch(DatabaseException e){
+            ctx.redirect("/");
             ctx.attribute("message",e.getMessage());
-            ctx.render("index.html");
         }
     }
     public static void select_order(Context ctx,ConnectionPool connectionPool){
-        int selectedOrderId = Integer.parseInt(ctx.formParam("selected_order"));
-        System.out.println(selectedOrderId);
-        ctx.sessionAttribute("selected_order",selectedOrderId);
-        ctx.render("admin.html");
+        try {
+            checkUsers(ctx, connectionPool);
+            if(ctx.formParam("selected_order")!=null) {
+                int selectedOrderId = Integer.parseInt(ctx.formParam("selected_order"));
+                System.out.println(selectedOrderId);
+                ctx.sessionAttribute("selected_order", selectedOrderId);
+            }
+        } catch(DatabaseException e){
+            ctx.redirect("/");
+            ctx.attribute("message",e.getMessage());
+        }
     }
 
      public static void allOrders(Context ctx, ConnectionPool connectionPool){
